# [MRG] Improving `MinHash.remove_many(...)` performance