* This PR improve the performance of the [Python:MinHash.remove_many](https://github.com/dib-lab/sourmash/blob/516dc53b65727e8a1379b8dc2971eda158aedecf/src/sourmash/minhash.py#L307L309) - [Rust:kmerminhash_remove_many](https://github.com/dib-lab/sourmash/blob/08b8cabdd7839002171ae884ddd55a0bb07e7991/src/core/src/ffi/minhash.rs#L147-L163)
    * **Current behavior:** Copying a Python list of hashes to Rust, then apply removal.
    * **Expected new behavior:** Pass the MinHash object by reference, similar to [add_many](https://github.com/dib-lab/sourmash/blob/516dc53b65727e8a1379b8dc2971eda158aedecf/src/sourmash/minhash.py#L296L305) then perform the removal.
* This PR Fixes #1553, and it's related to Ref: #1552.

:warning: This PR will touch all the layers (Python, C header, Rust).

---
*First time contributing: [https://orcid.org/0000-0002-3419-4785](https://orcid.org/0000-0002-3419-4785)*