@mr-eyes 2021-06-07
I am having a problem running `make benchmark` or `tox -e asv` for the new written benchmark function but here are the results of profiling the new function.

### Benchmark results

#### Time Profiling

```
Timer unit: 1e-06 s

Total time: 3.96304 s
File: profile_remove_many.py
Function: old at line 15

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
    15                                           @profile
    16                                           def old():
    17         1    3962976.0 3962976.0    100.0      mh1.remove_many(l)
    18         1         67.0     67.0      0.0      assert len(mh1) == 0

Total time: 5.7e-05 s
File: profile_remove_many.py
Function: new at line 21

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
    21                                           @profile
    22                                           def new():
    23         1         50.0     50.0     87.7      mh2.remove_many(rm_mh)
    24         1          7.0      7.0     12.3      assert len(mh2) == 0
```

---

#### Memory profiling

```
Filename: profile_remove_many.py

Line #    Mem usage    Increment  Occurences   Line Contents
============================================================
    15 3909.449 MiB 3909.449 MiB           1   @profile
    16                                         def old():
    17 3909.656 MiB    0.207 MiB           1       mh1.remove_many(l)
    18 3909.656 MiB    0.000 MiB           1       assert len(mh1) == 0


Filename: profile_remove_many.py

Line #    Mem usage    Increment  Occurences   Line Contents
============================================================
    21 3909.656 MiB 3909.656 MiB           1   @profile
    22                                         def new():
    23 3909.656 MiB    0.000 MiB           1       mh2.remove_many(rm_mh)
    24 3909.656 MiB    0.000 MiB           1       assert len(mh2) == 0
```

---

#### Reproducing the profiling

**profile_remove_many.py**
```python
from sourmash import MinHash

l = list(range(100000000))

mh1 = MinHash(500, 21, track_abundance=False)
mh1.add_many(l)

mh2 = MinHash(500, 21, track_abundance=False)
mh2.add_many(l)

rm_mh = MinHash(500, 21, track_abundance=False)
rm_mh.add_many(l)


@profile
def old():
    mh1.remove_many(l)
    assert len(mh1) == 0


@profile
def new():
    mh2.remove_many(rm_mh)
    assert len(mh2) == 0

old()
new()
```

**reproduce**

```bash
# time
kernprof -l profile_remove_many.py
python -m line_profiler profile_remove_many.py.lprof

# memory
python -m memory_profiler profile_remove_many.py
```