@mr-eyes 2021-06-07
Another simple profiling focusing on time 

```
add_many(iterable) took: 362.20860481262207 ms
add_many(MinHash) took: 0.09489059448242188 ms
```

---

```python
from sourmash import MinHash
from time import time

l = list(range(int(10000000)))

mh1 = MinHash(500, 21, track_abundance=False)
mh1.add_many(l)

mh2 = MinHash(500, 21, track_abundance=False)
mh2.add_many(l)

rm_mh = MinHash(500, 21, track_abundance=False)
rm_mh.add_many(l)

def old():
    mh1.remove_many(l)
    assert len(mh1) == 0


def new():
    mh2.remove_many(rm_mh)
    assert len(mh2) == 0

t = time()
old()
print(f"add_many(iterable) took: {1000 * (time() - t)} ms")

t = time()
new()
print(f"add_many(MinHash) took: {1000 * (time() - t)} ms")
```