@luizirber 2021-06-07
> This looks good to me but we should wait for @luizirber if we can :)

Good? GOOD? This is awesome! It's five orders of magnitude faster! =]
(will actually look at the code now, but really liked the benchmarking, Mo!)