@luizirber 2021-06-07
For fixing the two checks:
- https://github.com/dib-lab/sourmash/pull/1571/checks?check_run_id=2759014261#step:4:11 needs a `cargo fmt` and commit changes;
- https://github.com/dib-lab/sourmash/pull/1571/checks?check_run_id=2759014381#step:6:6 needs `cbindgen` to be rerun, mostly because of formatting of `include/sourmash.h`. I'll add the change as a suggestion, just a sec